benchmark/tts/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.31.0
tqdm>=4.65.0

jiwer>=3.0.0

librosa>=0.10.0
soundfile>=0.12.0

matplotlib>=3.7.0
seaborn>=0.12.0

# NISQA-TTS for naturalness evaluation
# TODO: verify later
# nisqa>=0.3.0
